% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atm.R
\docType{data}
\name{atm}
\alias{atm}
\title{Monoisotopic masses of 24 commonly-seen elements}
\format{
A list with 24 numeric values:
}
\source{
\url{https://www.sisweb.com/referenc/source/exactmas.htm}
}
\usage{
atm
}
\description{
A list object containing the monoisotopic masses of electron, proton, neutron, 14N, 15N, 12C,
13C, 14C, 1H, 2H, 16O, 18O, 6Li, 7Li, 23Na, 39K, 41K, 35Cl, 37Cl, 24Mg, 25Mg, 26Mg, 32S, 34S.
}
\keyword{datasets}
