% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixmergedknot.R
\name{calcdir}
\alias{calcdir}
\title{Calculate the direction of each knot}
\usage{
calcdir(ps)
}
\arguments{
\item{ps}{data.table A signle knot with aggregated information.}
}
\value{
numeric A value indicating the trend of the knot. Ascending intensities have positive value, 
descending intensities have negeative value.
}
\description{
Calculate the direction of each knot
}
\seealso{
\code{\link{findknots}}
}
