% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credential_knots.R
\name{calcratio}
\alias{calcratio}
\title{Calculation of ratio of unlabeled/labeled (12C/13C) peak intensities.}
\usage{
calcratio(knot)
}
\arguments{
\item{knot}{a set of knots to be calculated. The data structure are the same as}
}
\value{
numeric Ratio of measured intensities
}
\description{
Calculation of ratio of unlabeled/labeled (12C/13C) peak intensities.
}
\seealso{
\code{\link{findknots}} \code{\link{findknots}}
}
