% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixmergedknot.R
\name{checkmerged}
\alias{checkmerged}
\title{Search and label if certain increase and decrease are found at the head and tail region respectively.}
\usage{
checkmerged(knot)
}
\arguments{
\item{knot}{data.table A single knot with all aggregated information}
}
\value{
numeric a vector of values (1 or 2) indicating whether features in the knot should be splited or not.
}
\description{
Search and label if certain increase and decrease are found at the head and tail region respectively.
}
\seealso{
\code{\link{findknots}}
}
