% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credential_knots.R
\name{credential}
\alias{credential}
\title{Core algorithm of credentialknots: Heuristic search for knots that satisfy credentialing filters}
\usage{
credential(knots, Knot_quipu, ppmwid, rtwid, factor, mpc, ratio, ratio.lim, maxdimer, cd, do.plot)
}
\arguments{
\item{knots}{data.table. A subset of knots with same charge state to be credentialed.}

\item{Knot_quipu}{data.table The initial index of knot-to-quipu assignment.}

\item{ppmwid}{numeric Mass error tolerance (ppm) of credentialed knots.}

\item{rtwid}{numeric Retention time tolerance (seconds) of credentialed knots.}

\item{factor}{numeric Height cutoff of hierarchical clustering knots by (meanr, rt)}

\item{mpc}{numeric Mass-per-carbon filter. A vector of two values setting minimum and maximum theorectical mass per carbon.}

\item{ratio}{numeric The mixing ratio of unlabled/labeled (12C/13C) samples.}

\item{ratio.lim}{numeric Ratio filter. A decimal number (0,1] controling ranges of acceptable deviation of 
unlabeled/labeled ratio from mixing ratio.}

\item{maxdimer}{numeric Maximumly allowed number of knots in each credentialed groups, The default value is 4.}

\item{cd}{numeric Unit mass difference between unlabeled and labeled atoms. Defalut is 13C-12C = 13.00335 - 12.}

\item{do.plot}{logical If set to TRUE, plot the credentialing scores of the knots}
}
\value{
\code{Knot_quipu} data.table The new index of knot-to-quipu assignment after credentialing.
}
\description{
This functions identify credentialed knots pairs by intensity ratio check, mass-per-carbon check,
direction check followed by graph-clustering.
}
\seealso{
\code{\link{credentialknots}} \code{\link{findknots}}
}
\keyword{credentialing,}
\keyword{metabolomics}
