% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credential_main.R
\name{credentialing}
\alias{credentialing}
\title{#' Credentialing main function}
\usage{
credentialing(peaktable1, peaktable2, ppm, rtwin, rtcom, ratio1= 1/1, ratio2 = 1/2, ratio_tol=0.1, ratios_tol = 0.2, 
cd=13.00335-12, charges= 1:4, mpc=c(12,120), maxnmer=4)
}
\arguments{
\item{peaktable1}{data.table Feature table corresponding to the first ratio conditions.}

\item{peaktable2}{data.table Feature table corresponding to the second ratio conditions.}

\item{ppm}{numeric Mass error tolerance for knot searching and credentialing.}

\item{rtwin}{numeric Retention time window for 1st round credentialing}

\item{rtcom}{numeric Retention time window for 2nd round credentialing}

\item{ratio1}{numeric Ratio of designated 12C/13C ratio in peaktable1, default set to 1/1}

\item{ratio2}{numeric Ratio of designated 12C/13C ratio in the first peaktable, by default set to 1/2}

\item{ratio_tol}{numeric A decimal that controls the ratio range to pass the 1st round credentialing, The default value is 0.1.}

\item{ratios_tol}{numeric A decimal that controls the ratio range to pass the 2nd round credentialing, The default value is 0.2.}

\item{cd}{numeric Unit mass difference among isotopologues. Defalut is 13C-12C = 13.00335 - 12.}

\item{charges}{numeric Possible charge states to be considered when searching isotope pairs. Default value is 1:4.}

\item{mpc}{numeric A vector of two values setting minimal and maximal theoretical mass per carbon to be considered. The default value is c(12,120)}

\item{maxnmer}{numeric The maximal number of knots to accept within each credentialed features, The default value is 4.}
}
\value{
A list of credentialed features.
}
\description{
Credentialing workflow consists of four major parts. The analysis starts from two feature tables including index, m/z, retention time and intensity values for each isotope conditions.
First, credentialing searches for 'knots' at all possible charge states in each feature table. A knot is defined as a set of features with defined retention time window and m/z ppm error.
Second, credentialing breaks the knots with multiple isotopologue patterns. Third, credentialing clusteres searches for credentialed knots 
based on the similarity, charge state, isotopologue pattern, mass per carbon violation, and peak intensity ratios. Fourth, credentialing matches credentialed 'quipu' from both isotope mixing conditions
by checking head and tail peak m/z, retention time and intensity. 1st round credentialing is from step 1 to step 3 and 2nd round credentialing is step 4. 
The output of credentialing is a list consisting of credentialed features and credentialed knots at different levels.
}
\details{
The main function of credentialing.
}
\keyword{credentialing,}
\keyword{metabolomics}
\keyword{untargeted}
