% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credential_quipu.R
\name{credentialquipu}
\alias{credentialquipu}
\title{Step 4 of Credentialing: Find Matched Quipus between Two Credentialing Groups}
\usage{
credentialquipu(credentialedknots1, credentialedknots2, ppm = 15, rtwin = 1, ratio_ratio = 2, 
ratio_ratio_tol = 0.5, tailmatch=T)
}
\arguments{
\item{credentialedknots1}{list The original output of credentialknots() for the first credentialing group.}

\item{credentialedknots2}{list The original output of credentialknots() for the second credentialing group.}

\item{ppm}{numeric Maximum mass error tolerance when matching head and tail peaks of two quipus.}

\item{rtwin}{numeric Maximum mean retention time difference (rtmean) when matching two quipus.}

\item{ratio_ratio}{numeric Ratio of unlabeled/labeled mixing ratios of first and second credentialed samples. The default is 2.}

\item{ratio_ratio_tol}{numeric Ratio filter. A decimal number (0,1] controling ranges of acceptable deviation of 
ratio1/ratio2 from ratio_ratio, The default value is 0.5.}

\item{tailmatch}{logical Whether or not to include tail peak m/z when matching quipus.}
}
\value{
list A list of three tables "quipu_index" and quipu_match". \code{quipu_index} data.table 
The index of quipu-to-quipu assignment between two credentialing conditions. \code{quipu_match} data.table 
Aligned and merged table of quipu1 and quipu2 according to quipu_index.
}
\description{
This function matches two set of quipus (credentialed knots) by their retention time cutoff, 
head&tail m/z, charge state, peak intensities ratio.
}
\keyword{credentialing,}
\keyword{credentialquipu,}
\keyword{metabolomics}
