% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_knots.R
\name{findknots}
\alias{findknots}
\title{Find knots and resolve charge state}
\usage{
findknots(features, .zs = 1:4, ppmwid = 4, rtwid = 1, cd = 13.00335
  - 12)
}
\arguments{
\item{features}{data.frame Rows are features. Columns are "cc"}

\item{.zs}{integer A vector containing the absolute value of the possible charges.}

\item{ppmwid}{numeric The maximum mass error in ppm.}

\item{rtwid}{numeric The maximum retention time difference in seconds.}

\item{cd}{numeric The mass spacing to search for (defaults to C13 - C12)}
}
\value{
list A list with values "cc_knot" and "knot".  \code{cc_knot} contains the features assignments to a knot. \code{knot} contains aggregate information about each knot.
}
\description{
Find knots and resolve charge state
}
\seealso{
\code{\link{findzknots}}
}
