% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixmergedknot.R
\name{fixmergedknots}
\alias{fixmergedknots}
\title{Step 2 of Credentialing: Split Merged knots with Multiple Isotopologue Patterns}
\usage{
fixmergedknots(Knot, features)
}
\arguments{
\item{Knot}{list The output of findknots() function.}

\item{features}{data.table Feature table corresponding to the Knot.}
}
\value{
list A list with identical data sturcture as the input \code{Knot}.
}
\description{
This function breaks knots with multiple isotopologue patterns for credentialing step 3. If certain
descending and ascending pattern are found at the beginning (head) and ending (tail) region of the konts, the knot
are broken half as two new knots.
}
\seealso{
\code{\link{findknots}} \code{\link{credentialing}}
}
\keyword{credentialing,}
\keyword{fixmergedknot,}
\keyword{metabolomics}
