% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getXCMSfeature.R
\name{getXCMSfeature}
\alias{getXCMSfeature}
\title{Extract peak-table from an xcmsSet object}
\usage{
getXCMSfeature(xs,intchoice=c("into","intb","maxo"),sampling=1, sampleclass = NULL, export = F)
}
\arguments{
\item{xs}{xcmsSet object}

\item{intchoice}{character the intensity options provided by xcms. Options: \code{into}(default) -raw peak area, \code{intb}-baseline corrected peak area, \code{maxo}- max MS peak height.}

\item{sampling}{numeric The method to extract peak intensity in each feature group. If 1, sampling mean intensities of all peaks. If 2, sampling the maximum intensity of all peaks.}

\item{sampleclass}{character Specify sample classes to be extracted. The default is NULL, where all sample classes in "phenoData" table of xcmsSet.}

\item{export}{TRUE/FALSE Specify if the output features table should be exported to csv files and statistics exported to pdf file.}
}
\value{
a list The output includes two sets of data.table objects. \code{sampleclass-featuretable} data.table The features table from xcmsSet.
\code{sampleclass-credTable} data.table The structured features table for credentialing analysis.
}
\description{
The function extracts features table from xcmsSet object and return two types of data.table, one for reference, one for credentialing input
}
\details{
The adaptive function to automatically extract features (after grouping) from xcmsSet object
}
\author{
Lingjue "Mike" Wang <wang.lingjue@wustl.edu>
}
\keyword{credential}
\keyword{getXCMSfeature}
\keyword{getxcmsfeature}
\keyword{xcms}
