% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixmergedknot.R
\name{knotstats}
\alias{knotstats}
\title{Re-aggregate information of each knot after splitting}
\usage{
knotstats(knot)
}
\arguments{
\item{knot}{data.table A signle knot with aggregated information.}
}
\value{
data.table Aggregated information of new knots.
}
\description{
Re-aggregate information of each knot after splitting
}
\seealso{
\code{\link{findknots}}
}
