% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpc.R
\docType{data}
\name{mpc}
\alias{mpc}
\title{Minimum and maximum mass-per-carbon values of metabolites in E.coli}
\format{
A data.frame with three variables and 320 entries.
\describe{
\item{carbons}{The number of carbons}
\item{min_mpc}{maximum mass-per-carbon among metabolites with designated carbon number}
\item{max_mpc}{minimum mass-per-carbon among metabolites with designated carbon number}
}
}
\source{
\url{http://ecmdb.ca/}
}
\usage{
mpc
}
\description{
The minimum and maximum mass-per-carbons are used to refine the isotopologue searching.
}
\keyword{datasets}
