% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.knots}
\alias{plot.knots}
\title{Diagnostic Plotting Function for Knot}
\usage{
plot.knots(.knot, knots, features, scales = c(1,1), ...)
}
\arguments{
\item{.knot}{numeric Knot index.}

\item{knots}{list Knots. The output of \code{\link{findknots}}}

\item{features}{numeric The maximum retention time difference in seconds.}

\item{scales}{numeric Vector containing the minimum and maximum allowable mass per carbon.}

\item{...}{Further arguments to be passed.}
}
\value{
a ggplot object
}
\description{
This function generate plots of knots with m/z
}
