% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotcredpeaks}
\alias{plotcredpeaks}
\title{Step 4 of Credentialing: Plotting All Credentialed Peaks}
\usage{
plotcredpeaks(Credentialedindex, Credentialedpeaks, cred1 = "1T1", 
cred2 = "1T2", filename = "credentialedpeaks.pdf", ...)
}
\arguments{
\item{Credentialedindex}{data.table The index table of quipu-to-quipu assignment. A output table from \code{\link{credentialquipu}} or \code{\link{credentialing}}}

\item{Credentialedpeaks}{data.table The peak table includes the aligned peaks of all credentialed groups. A output table from \code{\link{credential}}}

\item{cred1}{character plotting legend of ratio1.}

\item{cred2}{character plotting legend of ratio2.}

\item{filename}{character the file name of the export pdf file Include .}

\item{...}{Further arguments to be passed.}
}
\value{
This plotting function have no functional output but an exported pdf file under working directory.
}
\description{
This function generates pseudo MS spectrum of all credentialed peak
groups. Each plot includes two pseudo-MS spectrum with each centroid peak represent a
credentialed isotopologue (the i refers to the intensity from feature table). The red spectra (up) is the
credentialed peak group in first ratio condition, the blue spectra (down) is the credentialed
peak group in second ratio condition. The legend represent the quipu number corresponding to
the matched credentialed groups in first and second conditions.
}
\seealso{
\code{\link{findknots}}
}
\keyword{credential}
\keyword{plotcredpeaks}
