% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotknots}
\alias{plotknots}
\title{Diagnostic Plotting Function for Knot}
\usage{
plotknots(.knot, knots, features, scales = c(1,1), ...)
}
\arguments{
\item{.knot}{numeric Knot index.}

\item{knots}{list Knots. The output of findknots()}

\item{features}{numeric The maximum retention time difference in seconds.}

\item{scales}{numeric Vector containing the minimum and maximum allowable mass per carbon.}

\item{...}{Further arguments to be passed.}
}
\value{
a ggplot object A list with values "knot_quipu" and "quipu".  \code{knot_quipu} assigns feature knots to quipu - credentialed groups. \code{quipu} contains aggregate information about each credentialed group.
}
\description{
This function generate plots of knots with m/z
}
\seealso{
\code{\link{findknots}}
}
