% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credential_knots.R
\name{credentialknots}
\alias{credentialknots}
\title{Heuristic search for knots which satisfy credentialing filters}
\usage{
credentialknots(Knots, ppmwid = 9, rtwid = 1, mpc = c(12, 120),
  Ratio = 1/1, Ratio.lim = 0.1, maxnmer = 4, cd = 13.00335 - 12,
  .zs = 1:4)
}
\arguments{
\item{Knots}{data.frame Rows are knots.}

\item{ppmwid}{numeric The maximum mass error in ppm.}

\item{rtwid}{numeric The maximum retention time difference in seconds.}

\item{mpc}{numeric Vector containing the minimum and maximum allowable mass per carbon.}

\item{Ratio}{numeric The ratio of U12C/U13C.}

\item{Ratio.lim}{numeric The maximum deviation from that ratio.}

\item{maxnmer}{integer The biggest n-mer to search for.}

\item{cd}{numeric The mass spacing to search for (defaults to C13 - C12)}

\item{.zs}{integer The charge states to search for.}
}
\value{
list A list with values "knot_quipu" and "quipu".  \code{knot_quipu} assigns feature knots to quipu - credentialed groups. \code{quipu} contains aggregate information about each credentialed group.
}
\description{
Heuristic search for knots which satisfy credentialing filters
}
\seealso{
\code{\link{findknots}}
}
