% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_knots.R
\name{findknots}
\alias{findknots}
\title{Step 1 of Credentialing: Find Isotopic Knots and Resolve Charge State}
\usage{
findknots(features, .zs=1:4, ppmwid=4, rtwid = 1, cd = 13.00335-12)
}
\arguments{
\item{features}{data.table A feature table including at least four exclusive columns: "cc", "mz", "rt" and "i". Each row is a feature (peak).}

\item{.zs}{integer A vector indicating all possible charge states to be searched against.}

\item{ppmwid}{numeric Mass error tolerance in ppm.}

\item{rtwid}{numeric The maximum retention time difference allowed in seconds.}

\item{cd}{numeric Unit mass difference between unlabeled and labeled atoms. Used for mass residual calculation. Defalut is 13C-12C = 13.00335 - 12.}
}
\value{
list A list with two tables "cc_knot" and "knot".  \code{cc_knot} data.table The index of feature-to-knot 
assignment. \code{knot} data.table Aggregated information of all knots: knot#, meanr (mean mass residual), meanmz 
(mean m/z), basemz (lowest mz peak), rt (mean rt), maxi (maximum intensity), n (number of peaks), z (charge state), 
dir (relative of the knot from left to right, positive value for increasing, negative for decreasing).
}
\description{
This function searches for 'knot' from feature table based on retention time and mass residual analysis. 
A knot is a set of isotopologue-like features with defined m/z and retention time difference and same charge states.
Based on the fact that isotopologues have identical mass residual against \code{cd} = 13C-12C (1.00335), mass residual 
analysis can be used to cluster isotopologues. Mass residual of a feature is calculated as: (mz %% cd).
}
\seealso{
\code{\link{credentialing}}
}
\keyword{credentialing,}
\keyword{findknots,}
\keyword{metabolomics}
